pyproject.toml
[tool.poetry]
name = "pd-7-hw-27"
version = "0.1.0"
description = ""
authors = ["SergeyLubash <luba-sergei@yandex.ru>"]
readme = "README.md"
packages = [{include = "pd_7_hw_27"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
pillow = "^9.2.0"
psycopg2 = "^2.9.5"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
